+++ test/Candoumbe.DataAccess.EFStore.UnitTests/Entities/Acolyte.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using Optional.Collections;
+using Ultimately.Collections;
 
 public class Acolyte
 {