+++ test/Candoumbe.DataAccess.EFStore.UnitTests/Entities/Hero.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
-using Optional.Collections;
+using Ultimately.Collections;
 
 public class Hero
 {