+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 💥 Breaking changes
+- Replaced [Optional](https://github.com/nlkl/Optional) dependency with [Ultimately](https://github.com/silkfire/ultimately) 
+
+
 
 ## [0.1.0] / 2025-09-17
 - Initial release